add error boundary decorator
