+++ packages/examples/.storybook/preview.ts
@@ -1,5 +1,7 @@
 import type { Preview } from "@storybook/react";
-
-const preview: Preview = {};
+import { withErrorBoundary } from "../src/testing-utils/deorators";
+const preview: Preview = {
+  decorators: [withErrorBoundary],
+};
 
 export default preview;
