+++ packages/examples/src/relay/Feedback/FeedbackContainer.tsx
@@ -2,6 +2,7 @@ import { graphql, useLazyLoadQuery } from "@nova/react";
 import { FeedbackComponent } from "./Feedback";
 import * as React from "react";
 import type { FeedbackContainerQuery as QueryType } from "./__generated__/FeedbackContainerQuery.graphql";
+import { ErrorBoundary } from "../../shared/ErrorBoundary";
 
 const FeedbackContainerQuery = graphql`
   query FeedbackContainerQuery($id: ID!) {
@@ -21,34 +22,6 @@ const FeedbackContainerInner = () => {
     return <div>No data</div>;
   }
 };
-
-class ErrorBoundary extends React.Component<
-  {
-    children: React.ReactNode;
-    onError?: (error: Error) => void;
-    fallback?: React.ReactNode;
-  },
-  { hasError: boolean }
-> {
-  constructor(props: { children: React.ReactNode }) {
-    super(props);
-    this.state = { hasError: false };
-  }
-  public componentDidCatch(error: Error): void {
-    this.props.onError?.(error);
-    this.setState({ hasError: true });
-  }
-
-  public static getDerivedStateFromError() {
-    return { hasError: true };
-  }
-  render() {
-    if (this.state.hasError) {
-      return <div>Error!</div>;
-    }
-    return this.props.children;
-  }
-}
 
 export const FeedbackContainer = () => {
   return (
